Refactor tests for Util\Crypto
